- Single-thread version of BlockSender, BlockServer.ReceiveBlock. - Fixed BinarySize comparisons in some unit tests.
git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/branches/singlethreaddfs@780 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ DataServer/ExtensionMethods.cs
@@ -13,7 +13,7 @@ namespace DataServerApplication
     {
         public static void WriteResult(this BinaryWriter writer, DataServerClientProtocolResult result)
         {
-            writer.Write((int)result);
+            writer.Write((short)result);
             writer.Flush();
         }
     }
